requirements.txt
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.6.0
bleach==5.0.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.0.4
colorama==0.4.4
cryptography==37.0.4
cycler==0.11.0
dateparser==1.1.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.5
django-cors-headers==3.13.0
django-jazzmin==2.5.0
django-jet==1.0.8
django-jet-reboot==1.3.1
django-rest-framework==0.1.0
django-simpleui==2022.6.131
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-typescript-generator==0.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.2
fastjsonschema==2.15.3
fonttools==4.34.4
fpdf==1.7.2
google-api-python-client==1.4.1
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
ipykernel==6.9.0
ipython==8.0.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.7.1
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.2.2
jupyterthemes==0.20.0
kiwisolver==1.4.3
lesscpy==0.15.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.4
notebook==6.4.12
numpy==1.23.0
oauth2client==4.1.3
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
platformdirs==2.5.2
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.27
psycopg2==2.8.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2022.3.2
reportlab==3.6.10
requests==2.28.1
rsa==4.8
selenium==4.3.0
Send2Trash==1.8.0
setuptools==61.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
stack-data==0.1.4
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
typed-ast==1.4.3
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
whitenoise==6.2.0
wincertstore==0.2
wsproto==1.1.0
